+++ client/src/container/serviceTypes.ts
@@ -1,7 +1,8 @@
 export const SERVICE_TYPES = {
   App: Symbol.for("App"),
+  KeyboardEvents: Symbol.for("KeyboardEvents"),
   LayoutService: Symbol.for("LayoutService"),
-  ResourceLoaderService: Symbol.for("ResourceLoaderService"),
+  ResourceFactoryBuilder: Symbol.for("ResourceFactoryBuilder"),
   Scene: Symbol.for("Scene"),
   StatsService: Symbol.for("StatsService"),
   WebGLRenderer: Symbol.for("WebGLRenderer")