+++ client/src/types/Scenario.ts
@@ -1,6 +1,4 @@
-import type { SceneComponent } from "./SceneComponent";
-
-export interface Scenario<TState extends string = string> {
-  start(sceneComponent: SceneComponent<TState>): void;
+export interface Scenario<TTarget = unknown> {
+  start(target: TTarget): void;
   stop(): void;
 }